fix: Convert MIN_SCENE_INTERVAL to int for ContentDetector to avoid errors

+++ drapto/video/scene_detection.py
@@ -33,7 +33,7 @@ def detect_scenes(input_file: Path) -> List[float]:
             log.debug("Starting scene detection on %s", input_file)
             scenes = detect(str(input_file),
                           ContentDetector(threshold=float(SCENE_THRESHOLD),
-                                        min_scene_len=float(MIN_SCENE_INTERVAL)))
+                                        min_scene_len=int(MIN_SCENE_INTERVAL)))
             log.debug("Raw scenes detected: %r", scenes)
         except Exception as e:
             log.error("Scene detection failed during detect() call: %s", e)
